+++ reports/planning_calendar/templates/ag7ifcal.go
@@ -31,25 +31,25 @@ const AG7IFTable = `\begin{center}
 \hline
     \textbf{QTR} & \textbf{XFR} & \textbf{MO} & \textbf{SP} & \multicolumn{2}{c|}{\textbf{WK}} & \textbf{START} & \textbf{END} \\
 \hline
-    \multirow{13}{*}{Q2} & \multirow{6}{*}{X3}  & \multirow{4}{*}{Q2-1} & \multirow{2}{*}{S07} & W14 &  A$\clubsuit$ & +WKSD14 & +WKED14 \\
-                         &                      &                       &                      & W15 &  2$\clubsuit$ & +WKSD15 & +WKED15 \\
+    \multirow{13}{*}{Q2} & \multirow{6}{*}{X3}  & \multirow{4}{*}{Q2-1} & \multirow{2}{*}{S07} & W14 &  A$\diamondsuit$ & +WKSD14 & +WKED14 \\
+                         &                      &                       &                      & W15 &  2$\diamondsuit$ & +WKSD15 & +WKED15 \\
 \cline{4-8}                                                                                                                             
-                         &                      &                       & \multirow{2}{*}{S08} & W16 &  3$\clubsuit$ & +WKSD16 & +WKED16 \\
-                         &                      &                       &                      & W17 &  4$\clubsuit$ & +WKSD17 & +WKED17 \\
+                         &                      &                       & \multirow{2}{*}{S08} & W16 &  3$\diamondsuit$ & +WKSD16 & +WKED16 \\
+                         &                      &                       &                      & W17 &  4$\diamondsuit$ & +WKSD17 & +WKED17 \\
 \cline{3-8}                                                                                                                             
-                         &                      & \multirow{4}{*}{Q2-2} & \multirow{2}{*}{S09} & W18 &  5$\clubsuit$ & +WKSD18 & +WKED18 \\
-                         &                      &                       &                      & W19 &  6$\clubsuit$ & +WKSD19 & +WKED19 \\
+                         &                      & \multirow{4}{*}{Q2-2} & \multirow{2}{*}{S09} & W18 &  5$\diamondsuit$ & +WKSD18 & +WKED18 \\
+                         &                      &                       &                      & W19 &  6$\diamondsuit$ & +WKSD19 & +WKED19 \\
 \cline{2-2}\cline{4-8}                                                                                                                  
-                         & \multirow{6}{*}{X4}  &                       & \multirow{2}{*}{S10} & W20 &  7$\clubsuit$ & +WKSD20 & +WKED20 \\
-                         &                      &                       &                      & W21 &  8$\clubsuit$ & +WKSD21 & +WKED21 \\
+                         & \multirow{6}{*}{X4}  &                       & \multirow{2}{*}{S10} & W20 &  7$\diamondsuit$ & +WKSD20 & +WKED20 \\
+                         &                      &                       &                      & W21 &  8$\diamondsuit$ & +WKSD21 & +WKED21 \\
 \cline{3-8}                                                                                                                             
-                         &                      & \multirow{5}{*}{Q2-3} & \multirow{2}{*}{S11} & W22 &  9$\clubsuit$ & +WKSD22 & +WKED22 \\
-                         &                      &                       &                      & W23 & 10$\clubsuit$ & +WKSD23 & +WKED23 \\
+                         &                      & \multirow{5}{*}{Q2-3} & \multirow{2}{*}{S11} & W22 &  9$\diamondsuit$ & +WKSD22 & +WKED22 \\
+                         &                      &                       &                      & W23 & 10$\diamondsuit$ & +WKSD23 & +WKED23 \\
 \cline{4-8}                                                                                                                             
-                         &                      &                       & \multirow{2}{*}{S12} & W24 &  J$\clubsuit$ & +WKSD24 & +WKED24 \\
-                         &                      &                       &                      & W25 &  Q$\clubsuit$ & +WKSD25 & +WKED25 \\
+                         &                      &                       & \multirow{2}{*}{S12} & W24 &  J$\diamondsuit$ & +WKSD24 & +WKED24 \\
+                         &                      &                       &                      & W25 &  Q$\diamondsuit$ & +WKSD25 & +WKED25 \\
 \cline{2-2}\cline{4-8}                                                                                                                  
-                         & SP2                  &                       & SP2                  & W26 &  K$\clubsuit$ & +WKSD26 & +WKED26 \\
+                         & SP2                  &                       & SP2                  & W26 &  K$\diamondsuit$ & +WKSD26 & +WKED26 \\
 \hline
 \end{tabular}
 
@@ -59,25 +59,25 @@ const AG7IFTable = `\begin{center}
 \hline
     \textbf{QTR} & \textbf{XFR} & \textbf{MO} & \textbf{SP} & \multicolumn{2}{c|}{\textbf{WK}} & \textbf{START} & \textbf{END} \\
 \hline
-    \multirow{13}{*}{Q3} & \multirow{6}{*}{X5}  & \multirow{4}{*}{Q3-1} & \multirow{2}{*}{S13} & W27 &  K$\diamondsuit$ & +WKSD27 & +WKED27 \\
-                         &                      &                       &                      & W28 &  Q$\diamondsuit$ & +WKSD28 & +WKED28 \\
+    \multirow{13}{*}{Q3} & \multirow{6}{*}{X5}  & \multirow{4}{*}{Q3-1} & \multirow{2}{*}{S13} & W27 &  A$\clubsuit$ & +WKSD27 & +WKED27 \\
+                         &                      &                       &                      & W28 &  2$\clubsuit$ & +WKSD28 & +WKED28 \\
 \cline{4-8}                                                                                                                                
-                         &                      &                       & \multirow{2}{*}{S14} & W29 &  J$\diamondsuit$ & +WKSD29 & +WKED29 \\
-                         &                      &                       &                      & W30 & 10$\diamondsuit$ & +WKSD30 & +WKED30 \\
+                         &                      &                       & \multirow{2}{*}{S14} & W29 &  3$\clubsuit$ & +WKSD29 & +WKED29 \\
+                         &                      &                       &                      & W30 &  4$\clubsuit$ & +WKSD30 & +WKED30 \\
 \cline{3-8}                                                                                                                                
-                         &                      & \multirow{4}{*}{Q3-2} & \multirow{2}{*}{S15} & W31 &  9$\diamondsuit$ & +WKSD31 & +WKED31 \\
-                         &                      &                       &                      & W32 &  8$\diamondsuit$ & +WKSD32 & +WKED32 \\
+                         &                      & \multirow{4}{*}{Q3-2} & \multirow{2}{*}{S15} & W31 &  5$\clubsuit$ & +WKSD31 & +WKED31 \\
+                         &                      &                       &                      & W32 &  6$\clubsuit$ & +WKSD32 & +WKED32 \\
 \cline{2-2}\cline{4-8}                                                                                                                     
-                         & \multirow{6}{*}{X6}  &                       & \multirow{2}{*}{S16} & W33 &  7$\diamondsuit$ & +WKSD33 & +WKED33 \\
-                         &                      &                       &                      & W34 &  6$\diamondsuit$ & +WKSD34 & +WKED34 \\
+                         & \multirow{6}{*}{X6}  &                       & \multirow{2}{*}{S16} & W33 &  7$\clubsuit$ & +WKSD33 & +WKED33 \\
+                         &                      &                       &                      & W34 &  8$\clubsuit$ & +WKSD34 & +WKED34 \\
 \cline{3-8}                                                                                                                                
-                         &                      & \multirow{5}{*}{Q3-3} & \multirow{2}{*}{S17} & W35 &  5$\diamondsuit$ & +WKSD35 & +WKED35 \\
-                         &                      &                       &                      & W36 &  4$\diamondsuit$ & +WKSD36 & +WKED36 \\
+                         &                      & \multirow{5}{*}{Q3-3} & \multirow{2}{*}{S17} & W35 &  9$\clubsuit$ & +WKSD35 & +WKED35 \\
+                         &                      &                       &                      & W36 & 10$\clubsuit$ & +WKSD36 & +WKED36 \\
 \cline{4-8}                                                                                                                                
-                         &                      &                       & \multirow{2}{*}{S18} & W37 &  3$\diamondsuit$ & +WKSD37 & +WKED37 \\
-                         &                      &                       &                      & W38 &  2$\diamondsuit$ & +WKSD38 & +WKED38 \\
+                         &                      &                       & \multirow{2}{*}{S18} & W37 &  J$\clubsuit$ & +WKSD37 & +WKED37 \\
+                         &                      &                       &                      & W38 &  Q$\clubsuit$ & +WKSD38 & +WKED38 \\
 \cline{2-2}\cline{4-8}                                                                                                                     
-                         & SP3                  &                       & SP3                  & W39 &  A$\diamondsuit$ & +WKSD39 & +WKED39 \\
+                         & SP3                  &                       & SP3                  & W39 &  K$\clubsuit$ & +WKSD39 & +WKED39 \\
 \hline
 \end{tabular}
 \quad
@@ -85,25 +85,25 @@ const AG7IFTable = `\begin{center}
 \hline
     \textbf{QTR} & \textbf{XFR} & \textbf{MO} & \textbf{SP} & \multicolumn{2}{c|}{\textbf{WK}} & \textbf{START} & \textbf{END} \\
 \hline
-    \multirow{13}{*}{Q4} & \multirow{6}{*}{X7}  & \multirow{4}{*}{Q4-1} & \multirow{2}{*}{S19} & W40 &  K$\spadesuit$ & +WKSD40 & +WKED40 \\
-                         &                      &                       &                      & W41 &  Q$\spadesuit$ & +WKSD41 & +WKED41 \\
+    \multirow{13}{*}{Q4} & \multirow{6}{*}{X7}  & \multirow{4}{*}{Q4-1} & \multirow{2}{*}{S19} & W40 &  A$\spadesuit$ & +WKSD40 & +WKED40 \\
+                         &                      &                       &                      & W41 &  2$\spadesuit$ & +WKSD41 & +WKED41 \\
 \cline{4-8}
-                         &                      &                       & \multirow{2}{*}{S20} & W42 &  J$\spadesuit$ & +WKSD42 & +WKED42 \\
-                         &                      &                       &                      & W43 & 10$\spadesuit$ & +WKSD43 & +WKED43 \\
+                         &                      &                       & \multirow{2}{*}{S20} & W42 &  3$\spadesuit$ & +WKSD42 & +WKED42 \\
+                         &                      &                       &                      & W43 &  4$\spadesuit$ & +WKSD43 & +WKED43 \\
 \cline{3-8}
-                         &                      & \multirow{4}{*}{Q4-2} & \multirow{2}{*}{S21} & W44 &  9$\spadesuit$ & +WKSD44 & +WKED44 \\
-                         &                      &                       &                      & W45 &  8$\spadesuit$ & +WKSD45 & +WKED45 \\
+                         &                      & \multirow{4}{*}{Q4-2} & \multirow{2}{*}{S21} & W44 &  5$\spadesuit$ & +WKSD44 & +WKED44 \\
+                         &                      &                       &                      & W45 &  6$\spadesuit$ & +WKSD45 & +WKED45 \\
 \cline{2-2}\cline{4-8}
                          & \multirow{6}{*}{X8}  &                       & \multirow{2}{*}{S22} & W46 &  7$\spadesuit$ & +WKSD46 & +WKED46 \\
-                         &                      &                       &                      & W47 &  6$\spadesuit$ & +WKSD47 & +WKED47 \\
+                         &                      &                       &                      & W47 &  8$\spadesuit$ & +WKSD47 & +WKED47 \\
 \cline{3-8}
-                         &                      & \multirow{5}{*}{Q4-3} & \multirow{2}{*}{S23} & W48 &  5$\spadesuit$ & +WKSD48 & +WKED48 \\
-                         &                      &                       &                      & W49 &  4$\spadesuit$ & +WKSD49 & +WKED49 \\
+                         &                      & \multirow{5}{*}{Q4-3} & \multirow{2}{*}{S23} & W48 &  9$\spadesuit$ & +WKSD48 & +WKED48 \\
+                         &                      &                       &                      & W49 & 10$\spadesuit$ & +WKSD49 & +WKED49 \\
 \cline{4-8}
-                         &                      &                       & \multirow{2}{*}{S24} & W50 &  3$\spadesuit$ & +WKSD50 & +WKED50 \\
-                         &                      &                       &                      & W51 &  2$\spadesuit$ & +WKSD51 & +WKED51 \\
+                         &                      &                       & \multirow{2}{*}{S24} & W50 &  J$\spadesuit$ & +WKSD50 & +WKED50 \\
+                         &                      &                       &                      & W51 &  Q$\spadesuit$ & +WKSD51 & +WKED51 \\
 \cline{2-2}\cline{4-8}
-                         & SP4                  &                       & SP4                  & W52 &  A$\spadesuit$ & +WKSD52 & +WKED52 \\
+                         & SP4                  &                       & SP4                  & W52 &  K$\spadesuit$ & +WKSD52 & +WKED52 \\
 \hline
 \end{tabular}
 \end{center}